Extract ComponentLimits to ui.constant
Move the ComponentLimits class from discord/ui/core.py into a new discord/ui/constant.py and update all UI modules to import ComponentLimits from .constant. Also export the constant module in discord/ui/__init__.py and remove ComponentLimits from core's exports. This centralizes UI constraint constants into a dedicated module.
