+++ discord/ui/__init__.py
@@ -28,3 +28,4 @@ from .separator import *
 from .text_display import *
 from .thumbnail import *
 from .view import *
+from .constant import *
