+++ discord/ui/checkbox.py
@@ -29,7 +29,7 @@ from typing import TYPE_CHECKING
 
 from ..components import Checkbox as CheckboxComponent
 from ..enums import ComponentType
-from .core import ComponentLimits
+from .constant import ComponentLimits
 from .item import ModalItem
 
 __all__ = ("Checkbox",)
