+++ discord/ui/file_upload.py
@@ -30,7 +30,7 @@ from typing import TYPE_CHECKING
 from ..components import FileUpload as FileUploadComponent
 from ..enums import ComponentType
 from ..message import Attachment
-from .core import ComponentLimits
+from .constant import ComponentLimits
 from .item import ModalItem
 
 __all__ = ("FileUpload",)
