+++ discord/ui/input_text.py
@@ -29,7 +29,7 @@ from typing import TYPE_CHECKING
 
 from ..components import InputText as InputTextComponent
 from ..enums import ComponentType, InputTextStyle
-from .core import ComponentLimits
+from .constant import ComponentLimits
 from .item import ModalItem
 
 __all__ = ("InputText", "TextInput")
