+++ discord/ui/radio_group.py
@@ -33,7 +33,7 @@ from ..components import RadioGroup as RadioGroupComponent
 from ..components import RadioGroupOption
 from ..enums import ComponentType
 from ..utils import MISSING
-from .core import ComponentLimits
+from .constant import ComponentLimits
 from .item import ModalItem
 
 __all__ = ("RadioGroup",)
